+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+=======
+## [Unreleased]
+
+### Fixed
+- Fix default priority downlink policy to update default preference correctly.
+
 ## [2.16.0] - 2021-08-17
 
 ### Added